fixed bugs in deeplink functionality
